% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_datasets.R
\name{sim_data}
\alias{sim_data}
\title{simulate independent spatial and temporal data}
\usage{
sim_data(time_model, postcode_pop, timespan = range(time_model$data$Date))
}
\arguments{
\item{time_model}{a poisson GLM for the temporal model}

\item{postcode_pop}{data frame of postcode,pop columns}

\item{N}{number to simulate}
}
\value{
data frame of simulated case postcodes and times
}
\description{
sample independent spatial and temporal data
}
\details{
etc
}
\author{
Barry Rowlingson
}
